requirements.txt
streamlit>=1.35.0
pandas>=2.2.2
gspread>=6.0.2
google-auth>=2.30.0
python-dateutil>=2.9.0
pytz>=2024.1
